messing around with coordinate formatting

+++ app/src/main/java/com/company/strengthtracker/presentation/test_screen/graph_utils/DateFormatter.kt
@@ -5,14 +5,14 @@ import java.time.ZoneId
 
 public class DateFormatter {
 
-    fun dateFormatter(date: Long): Long {
+    fun dateFormatter(date: Long): Float{
         return Instant.ofEpochMilli(date)
             .atZone(ZoneId.systemDefault())
             .toLocalDate()
             .toString()
             .replace("-", "")
             .replaceRange(0, 1, "")
-            .toLong()
+            .toFloat()
     }
 
     fun dateBuilder(date: Long): String {
